refactor(TeacherProjectService): Extract RemoveNodeIdFromTransitionsService
