+++ src/app/services/moveNodesService.spec.ts
@@ -3,6 +3,7 @@ import { MoveNodesService } from '../../assets/wise5/services/moveNodesService';
 import { copy } from '../../assets/wise5/common/object/object';
 import demoProjectJSON_import from './sampleData/curriculum/Demo.project.json';
 import { TeacherProjectService } from '../../assets/wise5/services/teacherProjectService';
+import { RemoveNodeIdFromTransitionsService } from '../../assets/wise5/services/removeNodeIdFromTransitionsService';
 import { StudentTeacherCommonServicesModule } from '../student-teacher-common-services.module';
 import { provideHttpClientTesting } from '@angular/common/http/testing';
 import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
@@ -16,9 +17,15 @@ let teacherProjectService: TeacherProjectService;
 describe('MoveNodesService', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({
-    imports: [StudentTeacherCommonServicesModule],
-    providers: [MoveNodesService, TeacherProjectService, provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
-});
+      imports: [StudentTeacherCommonServicesModule],
+      providers: [
+        MoveNodesService,
+        RemoveNodeIdFromTransitionsService,
+        TeacherProjectService,
+        provideHttpClient(withInterceptorsFromDi()),
+        provideHttpClientTesting()
+      ]
+    });
     service = TestBed.inject(MoveNodesService);
     demoProjectJSON = copy(demoProjectJSON_import);
     teacherProjectService = TestBed.inject(TeacherProjectService);
