+++ src/assets/wise5/authoringTool/project-authoring-step/project-authoring-step.component.spec.ts
@@ -17,6 +17,7 @@ import { DeleteTranslationsService } from '../../services/deleteTranslationsServ
 import { provideRouter } from '@angular/router';
 import { CopyTranslationsService } from '../../services/copyTranslationsService';
 import { TeacherProjectTranslationService } from '../../services/teacherProjectTranslationService';
+import { RemoveNodeIdFromTransitionsService } from '../../services/removeNodeIdFromTransitionsService';
 import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 const nodeId1 = 'nodeId1';
@@ -44,6 +45,7 @@ describe('ProjectAuthoringStepComponent', () => {
         DeleteNodeService,
         DeleteTranslationsService,
         provideRouter([]),
+        RemoveNodeIdFromTransitionsService,
         TeacherDataService,
         TeacherProjectService,
         TeacherProjectTranslationService,
